rework issuance tests
