+++ test/integration/transaction.spec.ts
@@ -1,4 +1,3 @@
-import { IssuanceBlindingKeys } from './../../types/types.d';
 import * as assert from 'assert';
 import * as bip32 from 'bip32';
 import * as liquid from '../..';
@@ -1590,129 +1589,9 @@ describe('liquidjs-lib (transactions with psbt)', () => {
     const hex = psbt.extractTransaction().toHex();
     await regtestUtils.broadcast(hex);
   });
-
-  it('can create a 1-to-1 confidential Transaction (and broadcast via 3PBP) with unblinded issuance', async () => {
-    const blindingPrivkeys = [
-      Buffer.from(
-        '13d4dbfdb5074705e6b9758d1542d7dd8c03055086c0da421620eaa04717a9f7',
-        'hex',
-      ),
-    ];
-    const blindingPubkeys = ['', '', ''].map(
-      () => liquid.ECPair.makeRandom({ network: regtest }).publicKey,
-    );
-
-    const psbt = new liquid.Psbt();
-    psbt.setVersion(2); // These are defaults. This line is not needed.
-    psbt.setLocktime(0); // These are defaults. This line is not needed.
-    psbt.addInput({
-      // if hash is string, txid, if hash is Buffer, is reversed compared to txid
-      hash: 'dd983c9c0419fce6bcc0eaf875b54a2c19f9d6e761faa58b1afd199638275475',
-      index: 0,
-      // non-segwit inputs now require passing the whole previous tx as Buffer
-      nonWitnessUtxo: nonWitnessUtxoBuffer,
-    });
-    psbt.addIssuance({
-      assetAddress: 'XBXiDkFNneyPtpXvqVWQoHA1MhoXa8FZLn',
-      assetAmount: 100,
-      tokenAddress: 'XBXiDkFNneyPtpXvqVWQoHA1MhoXa8FZLn',
-      tokenAmount: 1,
-      precision: 8,
-      net: regtest,
-    });
-    psbt.addOutputs([
-      {
-        nonce,
-        asset,
-        value: liquid.confidential.satoshiToConfidentialValue(99996500),
-        script: Buffer.from(
-          '76a914659bedb5d3d3c7ab12d7f85323c3a1b6c060efbe88ac',
-          'hex',
-        ),
-      },
-      {
-        nonce,
-        asset,
-        value: liquid.confidential.satoshiToConfidentialValue(3500),
-        script: Buffer.alloc(0),
-      },
-    ]);
-    await psbt.blindOutputs(blindingPrivkeys, blindingPubkeys);
-    psbt.signInput(0, bob);
-    psbt.validateSignaturesOfInput(0);
-    psbt.finalizeAllInputs();
-    const hex = psbt.extractTransaction().toHex();
-    assert.doesNotReject(() => regtestUtils.broadcast(hex));
-  });
-
-  it('can create a 1-to-1 confidential Transaction (and broadcast via 3PBP) with blinded issuance', async () => {
-    const blindingPrivkeys = [
-      Buffer.from(
-        '13d4dbfdb5074705e6b9758d1542d7dd8c03055086c0da421620eaa04717a9f7',
-        'hex',
-      ),
-    ];
-    const blindingPubkeys = ['', '', ''].map(
-      () => liquid.ECPair.makeRandom({ network: regtest }).publicKey,
-    );
-
-    const issuanceBlindingKeys = ['', ''].map(
-      () => liquid.ECPair.makeRandom({ network: regtest }).privateKey!,
-    );
-
-    const psbt = new liquid.Psbt();
-    psbt.setVersion(2); // These are defaults. This line is not needed.
-    psbt.setLocktime(0); // These are defaults. This line is not needed.
-    psbt.addInput({
-      // if hash is string, txid, if hash is Buffer, is reversed compared to txid
-      hash: 'dd983c9c0419fce6bcc0eaf875b54a2c19f9d6e761faa58b1afd199638275475',
-      index: 0,
-      // non-segwit inputs now require passing the whole previous tx as Buffer
-      nonWitnessUtxo: nonWitnessUtxoBuffer,
-    });
-    psbt.addIssuance({
-      assetAddress: 'XBXiDkFNneyPtpXvqVWQoHA1MhoXa8FZLn',
-      assetAmount: 100,
-      tokenAddress: 'XBXiDkFNneyPtpXvqVWQoHA1MhoXa8FZLn',
-      tokenAmount: 1,
-      precision: 8,
-      net: regtest,
-    });
-    psbt.addOutputs([
-      {
-        nonce,
-        asset,
-        value: liquid.confidential.satoshiToConfidentialValue(99996500),
-        script: Buffer.from(
-          '76a914659bedb5d3d3c7ab12d7f85323c3a1b6c060efbe88ac',
-          'hex',
-        ),
-      },
-      {
-        nonce,
-        asset,
-        value: liquid.confidential.satoshiToConfidentialValue(3500),
-        script: Buffer.alloc(0),
-      },
-    ]);
-    await psbt.blindOutputsByIndex(
-      new Map<number, Buffer>().set(0, blindingPrivkeys[0]),
-      new Map<number, Buffer>()
-        .set(2, blindingPubkeys[2]),
-      new Map<number, IssuanceBlindingKeys>().set(0, {
-        assetKey: issuanceBlindingKeys[0],
-        tokenKey: issuanceBlindingKeys[1],
-      }),
-    );
-    psbt.signInput(0, bob);
-    psbt.validateSignaturesOfInput(0);
-    psbt.finalizeAllInputs();
-    const hex = psbt.extractTransaction().toHex();
-    assert.doesNotReject(() => regtestUtils.broadcast(hex));
-  });
 });
 
-function createPayment(
+export function createPayment(
   _type: string,
   myKeys?: any[],
   network?: any,
@@ -1789,7 +1668,7 @@ function getAddress(script: any, scriptType: string): string {
   throw new Error('Invalid script type');
 }
 
-async function getInputData(
+export async function getInputData(
   payment: any,
   isSegwit: boolean,
   redeemType: string,
@@ -1830,7 +1709,7 @@ async function getInputData(
   };
 }
 
-const nonWitnessUtxoBuffer = Buffer.from(
+export const nonWitnessUtxoBuffer = Buffer.from(
   '020000000101bb6d18772599ae3ad9c44c524d63b666747c1c195c6f516a41a8d5f4a' +
   '32eef05010000001716001408a2534a4b37e12c371886ead381981413eec5edfdffff' +
   'ff030b0cbc7820f47ecff027603c63c52f715884852f2bebbbf66e7cedb6de6682f3c' +
